// src/_store/types/worktime.types.ts
export interface IScheduleItem {
  wiId?: string;
  idWorkTemplate?: string;
  /** id дня недели*/
  dayId: string;
  /** начало работы в минутах 9:00 = 9*60 = 540*/
  start: number;
  /** конец работы в минутах*/
  end: number;
  /** доступно ли поле для изменения*/
  editable?: boolean;
}
interface IFlags {
  wiId?: string;
  field?: string;
  editable: boolean;
  show: boolean;
  required: boolean;
}
export interface IScheduleItemForm extends IScheduleItem {
  endFlags: IFlags;
  startFlags: IFlags;
}

export interface IBasicWorkTime {
  wiId?: string;
  /** id рабочего графика из справочника*/
  idWorkTemplate: string;
  /** имя рабочего графика из справочника*/
  nameWorkTemplate: string;
  /** массив деталей рабочего времени по дням недели*/
  workTime: IScheduleItem[];
  /** начало действия расписания*/
  startDate?: string;
  /** окончание действия расписания*/
  endDate?: string;
  /** число рабочего времени в минутах*/
  timePerWeek: number;
  /** Вид перерыва на кормление (1 - 30 мин. / 2 - 60 мин. / 3 - 30/60 мин.)*/
  breakFeedType?: string;
  /** Вид перерыва на кормление (описание)*/
  breakFeedTypeDesc?: string;
  /** Период перерыва на кормление (1 - Суммарно в начале дня / 2 - Суммарно к обеденному перерыву / 3 - Суммарно в конце дня / 4 - Каждые 3 часа*/
  breakFeedPer?: string;
  /** Период перерыва на кормление (описание)*/
  breakFeedPerDesc?: string;
  /** перерыв в минутах*/
  break: number;
  /** процент рабочего времени*/
  percentage?: number;
  /** комментарий*/
  comment?: string;
  /** существующая в системе заявка*/
  createdTask?: {
    procGuid: string;
    wiId: string;
    procEmpId: string;
    referenceNumber: string;
    procStatus: string;
    scenarioStage: string;
    stepStatus: string;
  };
}

export interface IBasicWorkTimeForm extends IBasicWorkTime {
  workTemplateFlags: IFlags;
  startDateFlags: IFlags;
  endDateFlags: IFlags;
  breakFlags: IFlags;
  workTime: IScheduleItemForm[];
}